+++ taipei-day-trip/TaipeiTravel/static/booking.js
