+++ taipei-day-trip/TaipeiTravel/static/image/attraction-leftArrow.png
