+++ taipei-day-trip/TaipeiTravel/static/image/mrtbar-leftArrow-hover.png
